Limit notification period when a fast track denied
If client requests a fast track and this request is denied, then don't
allow notification period to exceed FMS_20.  FMS_20 is the minimum task
wakeup period in ms for which CFS operates reliably.  This reduces
glitching when an app uses a buffer size that is too small for the endpoint.
The theory is that if the app wanted a fast track but couldn't get it,
then at least wake it up as often as possible (while still not waking
up so often than CFS can't handle it).

Also clarify in the log when the denial is by client vs server.

Test: as described in bug
Bug: 63040475
Change-Id: I2c171bda7929a8f309c2286d868e09887680d234
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>